+++ Assets/Scripts/Level/Entities/LivingEntity/Targeter/PlayerMouseTargeter.cs
@@ -7,13 +7,13 @@ public class PlayerMouseTargeter : TargeterBase {
 
     public PlayerMouseTargeter(LivingEntity owner)
         : base(owner) {
-        EventManager.StartListening((int) GameEvents.Mouse_Left_Press, SelectTarget);
-        EventManager.StartListening((int) GameEvents.Mouse_Left_Held, UpdateMoveLocation);
+        EventManager.StartListening(GameEvents.Mouse_Left_Press, SelectTarget);
+        EventManager.StartListening(GameEvents.Mouse_Left_Held, UpdateMoveLocation);
     }
 
     ~PlayerMouseTargeter() {
-        EventManager.StopListening((int) GameEvents.Mouse_Left_Press, SelectTarget);
-        EventManager.StopListening((int) GameEvents.Mouse_Left_Held, UpdateMoveLocation);
+        EventManager.StopListening(GameEvents.Mouse_Left_Press, SelectTarget);
+        EventManager.StopListening(GameEvents.Mouse_Left_Held, UpdateMoveLocation);
     }
 
     void SelectTarget(int param) {
@@ -28,12 +28,12 @@ public class PlayerMouseTargeter : TargeterBase {
                 hitType = EntityTypeToTargetType(hitEntity.GetEntity().GetEntityType());
                 targetedEntity = hitEntity;
                 targetedLocation = null;
-                EventManager.TriggerEvent(owner.GetEntityID(), (int) GameEvents.Targeted_Entity);
+                EventManager.TriggerEvent(owner.GetEntityID(), GameEvents.Targeted_Entity);
             } else {
                 hitType = TargetType.World;
                 targetedEntity = null;
                 targetedLocation = hit.point;
-                EventManager.TriggerEvent(owner.GetEntityID(), (int) GameEvents.Targeted_World);
+                EventManager.TriggerEvent(owner.GetEntityID(), GameEvents.Targeted_World);
             }
         } else {
             //  NOTE: [Rock]: There shouldn't really ever be an instance where we don't collide with something in the future, however we'll keep this here for now
@@ -69,7 +69,7 @@ public class PlayerMouseTargeter : TargeterBase {
             //  NOTE: [Rock]: In the future we'll probably want to have an ignore mask here...
             if (Physics.Raycast(ray, out hit, RAYCAST_DISTANCE)) {
                 targetedLocation = hit.point;
-                EventManager.TriggerEvent(owner.GetEntityID(), (int) GameEvents.Targeted_World);
+                EventManager.TriggerEvent(owner.GetEntityID(), GameEvents.Targeted_World);
             }
         }
     }
